Convert FreetextFilter to Typescript
format

format filter
